Format a summarised_result object into a gt, flextable or tibble object

formatTable() was renamed to visOmopTable()

Usage

formatTable(
  result,
  formatEstimateName,
  header,
  split,
  groupColumn = NULL,
  type = "gt",
  renameColumns = NULL,
  minCellCount = lifecycle::deprecated(),
  hide = c("result_id", "estimate_type"),
  .options = list()
)

Arguments

result

A summarised_result.

formatEstimateName

Named list of estimate name's to join, sorted by computation order. Indicate estimate_name's between <...>.

header

A vector containing which elements should go into the header in order (cdm_name, group, strata, additional, variable, estimate, and settings).

split

A vector containing the name-level groups to split ("group", "strata", "additional"), or an empty character vector to not split.

groupColumn

Columns to use as group labels. By default the name of the new group will be the column names separated by "_". To specify a new grouping name enter a named list such as: list(newGroupName = c("variable_name", "variable_level"))

type

Type of desired formatted table, possibilities: "gt", "flextable", "tibble".

renameColumns

Named vector to customise column names, for instance: c("Database name" = "cdm_name")). By default column names are transformed to sentence case.

minCellCount

Suppression of estimates when counts < minCellCount should be done before with ompogenerics::suppress().

hide

Columns to drop from the output table.

.options

Named list with additional formatting options. visOmopResults::optionsVisOmopTable() shows allowed arguments and their default values.

Value

A tibble, gt, or flextable object.

Details

Examples

mockSummarisedResult() |> formatTable(
  formatEstimateName = c("N%" = "<count> (<percentage>)",
                         "N" = "<count>",
                         "Mean (SD)" = "<mean> (<sd>)"),
  header = c("group"),
  split = c("group","strata",  "additional")
)
#> Warning: `formatTable()` was deprecated in visOmopResults 0.3.0.
#> ℹ Please use `visOmopTable()` instead.
#> ! Results have not been suppressed.

  

CDM name
      	Age group
      	Sex
      	Variable name
      	Variable level
      	Estimate name
      	
        Cohort name
      
    

Cohort1
      	Cohort2
    

mock
	Overall
	Overall
	Number subjects
	-
	N
	1,163,092
	4,414,992

	<40
	Male
	Number subjects
	-
	N
	7,154,231
	3,155,947

	>=40
	Male
	Number subjects
	-
	N
	2,261,087
	1,013,138

	<40
	Female
	Number subjects
	-
	N
	1,333,485
	2,726,462

	>=40
	Female
	Number subjects
	-
	N
	9,272,818
	6,536,994

	Overall
	Male
	Number subjects
	-
	N
	8,941,227
	9,278,701

	
	Female
	Number subjects
	-
	N
	2,036,644
	266,247

	<40
	Overall
	Number subjects
	-
	N
	2,569,829
	5,594,912

	>=40
	Overall
	Number subjects
	-
	N
	6,140,056
	8,541,737

	Overall
	Overall
	Age
	-
	Mean (SD)
	7.59 (5.24)
	72.82 (0.75)

	<40
	Male
	Age
	-
	Mean (SD)
	20.58 (4.97)
	97.08 (9.70)

	>=40
	Male
	Age
	-
	Mean (SD)
	49.59 (0.15)
	53.52 (6.49)

	<40
	Female
	Age
	-
	Mean (SD)
	26.19 (9.58)
	16.65 (9.85)

	>=40
	Female
	Age
	-
	Mean (SD)
	41.60 (4.53)
	5.82 (5.36)

	Overall
	Male
	Age
	-
	Mean (SD)
	43.91 (9.82)
	45.91 (7.61)

	
	Female
	Age
	-
	Mean (SD)
	18.24 (5.45)
	62.28 (5.64)

	<40
	Overall
	Age
	-
	Mean (SD)
	7.83 (3.60)
	59.48 (7.78)

	>=40
	Overall
	Age
	-
	Mean (SD)
	1.78 (5.03)
	74.30 (6.39)

	Overall
	Overall
	Medications
	Amoxiciline
	N%
	37,896 (92.15)
	82,792 (4.30)

	<40
	Male
	Medications
	Amoxiciline
	N%
	35,953 (67.57)
	64,307 (68.46)

	>=40
	Male
	Medications
	Amoxiciline
	N%
	84,423 (35.06)
	70,848 (51.44)

	<40
	Female
	Medications
	Amoxiciline
	N%
	45,675 (0.87)
	87,073 (94.57)

	>=40
	Female
	Medications
	Amoxiciline
	N%
	17,903 (43.09)
	28,181 (84.31)

	Overall
	Male
	Medications
	Amoxiciline
	N%
	96,621 (72.49)
	17,322 (6.49)

	
	Female
	Medications
	Amoxiciline
	N%
	71,341 (41.23)
	10,967 (76.21)

	<40
	Overall
	Medications
	Amoxiciline
	N%
	2,489 (55.09)
	58,622 (7.84)

	>=40
	Overall
	Medications
	Amoxiciline
	N%
	62,646 (15.64)
	85,072 (13.35)

	Overall
	Overall
	Medications
	Ibuprofen
	N%
	80,970 (73.11)
	68,080 (61.03)

	<40
	Male
	Medications
	Ibuprofen
	N%
	55,938 (85.24)
	85,929 (12.11)

	>=40
	Male
	Medications
	Ibuprofen
	N%
	73,373 (97.58)
	83,863 (62.94)

	<40
	Female
	Medications
	Ibuprofen
	N%
	34,970 (11.30)
	88,131 (71.20)

	>=40
	Female
	Medications
	Ibuprofen
	N%
	98,267 (97.03)
	23,770 (61.21)

	Overall
	Male
	Medications
	Ibuprofen
	N%
	40,885 (64.81)
	57,863 (3.44)

	
	Female
	Medications
	Ibuprofen
	N%
	36,087 (5.61)
	89,790 (66.59)

	<40
	Overall
	Medications
	Ibuprofen
	N%
	65,251 (47.31)
	34,443 (65.30)

	>=40
	Overall
	Medications
	Ibuprofen
	N%
	14,356 (29.46)
	76,086 (45.19)