Creates a gt object from a dataframe

Creates a flextable object from a dataframe using a delimiter to span the header, and allows to easily customise table style.

Usage

gtTable(
  x,
  delim = "\n",
  style = "default",
  na = "-",
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  groupColumn = NULL,
  groupNameCol = lifecycle::deprecated(),
  groupAsColumn = FALSE,
  groupNameAsColumn = lifecycle::deprecated(),
  groupOrder = NULL,
  colsToMergeRows = NULL
)

Arguments

x

A dataframe.

delim

Delimiter.

style

Named list that specifies how to style the different parts of the gt table. Accepted entries are: title, subtitle, header, header_name, header_level, column_name, group_label, and body. Alternatively, use "default" to get visOmopResults style, or NULL for gt style

na

How to display missing values.

title

Title of the table, or NULL for no title.

subtitle

Subtitle of the table, or NULL for no subtitle.

caption

Caption for the table, or NULL for no caption. Text in markdown formatting style (e.g. *Your caption here* for caption in italics).

groupColumn

Specifies the columns to use for group labels. By default, the new group name will be a combination of the column names, joined by "_". To assign a custom group name, provide a named list such as: list(newGroupName = c("variable_name", "variable_level"))

groupNameCol

groupAsColumn

Whether to display the group labels as a column (TRUE) or rows (FALSE).

groupNameAsColumn

groupOrder

Order in which to display group labels.

colsToMergeRows

Names of the columns to merge vertically when consecutive row cells have identical values. Alternatively, use "all_columns" to apply this merging to all columns, or use NULL to indicate no merging.

Value

gt object.

A gt table.

Examples

mockSummarisedResult() |>
  formatEstimateValue(decimals = c(integer = 0, numeric = 1)) |>
  formatHeader(header = c("Study strata", "strata_name", "strata_level"),
              includeHeaderName = FALSE) |>
  gtTable(
    style = list("header" = list(
      gt::cell_fill(color = "#d9d9d9"),
      gt::cell_text(weight = "bold")),
      "header_level" = list(gt::cell_fill(color = "#e1e1e1"),
                            gt::cell_text(weight = "bold")),
      "column_name" = list(gt::cell_text(weight = "bold")),
      "title" = list(gt::cell_text(weight = "bold"),
                     gt::cell_fill(color = "#c8c8c8")),
      "group_label" = gt::cell_fill(color = "#e1e1e1")),
    na = "--",
    title = "gtTable example",
    subtitle = NULL,
    caption = NULL,
    groupColumn = "group_level",
    groupAsColumn = FALSE,
    groupOrder = c("cohort1", "cohort2"),
    colsToMergeRows = "all_columns"
  )

  

gtTable example
    

      	
        Study strata
      
    

result_id
      	cdm_name
      	group_name
      	variable_name
      	variable_level
      	estimate_name
      	estimate_type
      	additional_name
      	additional_level
      	
        overall
      
      	
        age_group &&& sex
      
      	
        sex
      
      	
        age_group
      
    

overall
      	<40 &&& Male
      	>=40 &&& Male
      	<40 &&& Female
      	>=40 &&& Female
      	Male
      	Female
      	<40
      	>=40
    

cohort1
    

1
	mock
	cohort_name
	number subjects
	--
	count
	integer
	overall
	overall
	4,621,038
	9,392,060
	62,914
	3,494,275
	5,662,714
	5,289,571
	6,370,154
	3,507,959
	9,241,712

	
	
	age
	--
	mean
	numeric
	overall
	overall
	93.6
	67.2
	37.1
	56.7
	47.9
	56.8
	28.7
	76.0
	22.8

	
	
	
	
	sd
	numeric
	overall
	overall
	0.3
	7.4
	5.9
	6.9
	7.4
	4.6
	6.6
	0.0
	9.8

	
	
	Medications
	Amoxiciline
	count
	integer
	overall
	overall
	21,325
	79,201
	59,263
	60,626
	11,077
	19,561
	707
	94,087
	6,821

	
	
	
	
	percentage
	percentage
	overall
	overall
	19.4796363823116
	67.9479371290654
	74.7466302476823
	64.9537714431062
	7.75806771125644
	59.9653656128794
	87.9010812146589
	73.2037156121805
	8.00530861597508

	
	
	
	Ibuprofen
	count
	integer
	overall
	overall
	48,522
	47,024
	79,105
	31,298
	30,876
	34,516
	82,844
	27,400
	43,084

	
	
	
	
	percentage
	percentage
	overall
	overall
	12.5297799473628
	56.5557193476707
	12.7641106490046
	62.7597616286948
	7.53239539917558
	28.5451323958114
	31.0602429090068
	14.6727561717853
	79.9770513316616

cohort2
    

1
	mock
	cohort_name
	number subjects
	--
	count
	integer
	overall
	overall
	4,740,412
	7,781,341
	7,987,365
	8,784,999
	9,098,439
	4,455,378
	2,895,593
	3,033,107
	9,190,243

	
	
	age
	--
	mean
	numeric
	overall
	overall
	42.7
	45.7
	71.0
	28.7
	77.3
	57.5
	2.2
	77.3
	5.9

	
	
	
	
	sd
	numeric
	overall
	overall
	2.2
	3.1
	7.3
	1.3
	1.0
	1.7
	6.2
	6.8
	7.9

	
	
	Medications
	Amoxiciline
	count
	integer
	overall
	overall
	13,693
	33,282
	48,969
	17,295
	6,494
	45,167
	37,940
	90,930
	2,995

	
	
	
	
	percentage
	percentage
	overall
	overall
	83.2749811932445
	86.4705697866157
	71.2964971549809
	75.7407110882923
	12.2433916199952
	19.3697881884873
	31.4178289379925
	13.7532982043922
	84.7802880220115

	
	
	
	Ibuprofen
	count
	integer
	overall
	overall
	41,690
	13,431
	68,634
	17,208
	92,349
	61,374
	31,563
	80,973
	90,237

	
	
	
	
	percentage
	percentage
	overall
	overall
	17.1996079385281
	78.5683812340721
	47.0035764155909
	43.2957729091868
	57.4532859958708
	72.9969945270568
	89.4187148427591
	22.030636924319
	29.1596438270062