Helper for consistent documentation of plots.

Helper for consistent documentation of plots.

Arguments

result

A <summarised_result> object.

x

Column or estimate name that is used as x variable.

y

Column or estimate name that is used as y variable.

width

Bar width, as in geom_col() of the ggplot2 package.

just

Adjustment for column placement, as in geom_col() of the ggplot2 package.

facet

Variables to facet by, a formula can be provided to specify which variables should be used as rows and which ones as columns.

colour

Columns to use to determine the colours.

group

Columns to use to determine the group.

label

Character vector with the columns to display interactively in plotly.

style

A character string defining the visual theme to apply to the plot. You can set this to NULL to apply the standard ggplot2 default style, or provide a name for one of the package's pre-defined styles. Refer to the plotStyle() function for all available style pre-defined themes. For further customization, you can always modify the returned ggplot object directly.

type

The desired format of the output plot. See plotType() for supported plot types.

lower

Estimate name for the lower quantile of the box.

middle

Estimate name for the middle line of the box.

upper

Estimate name for the upper quantile of the box.

line

Whether to plot a line using geom_line.

point

Whether to plot points using geom_point.

ribbon

Whether to plot a ribbon using geom_ribbon.

ymin

Lower limit of error bars, if provided is plot using geom_errorbar.

ymax

Upper limit of error bars, if provided is plot using geom_errorbar.

position

Position of bars, can be either dodge or stack