Split name and level columns into the columns

Pivots the input dataframe so the values of the name columns are transformed into columns, which values come from the specified level column.

Usage

splitNameLevel(
  result,
  name = "group_name",
  level = "group_level",
  keep = FALSE,
  fill = "overall"
)

Arguments

result

A summarised_result object.

name

Column with the names.

level

Column with the levels.

keep

Whether to keep the original group_name and group_level columns.

fill

Optionally, a character that specifies what value should be filled in with when missing.

Value

A dataframe with the specified name column values as columns.