Format a summarised_result object into a gt, flextable, or tibble object.

This function combines the functionalities of formatEstimateValue(), formatEstimateName(), formatHeader(), gtTable(), and fxTable() into a single function specifically for summarised_result objects.

Usage

visOmopTable(
  result,
  formatEstimateName = character(),
  header = character(),
  settingsColumns = character(),
  groupColumn = character(),
  renameColumns = character(),
  type = "gt",
  hide = character(),
  showMinCellCount = TRUE,
  .options = list(),
  split = lifecycle::deprecated(),
  minCellCount = lifecycle::deprecated()
)

Arguments

result

A summarised_result object.

formatEstimateName

A named list of estimate names to join, sorted by computation order. Use <...> to indicate estimate names.

header

A vector specifying the elements to include in the header. The order of elements matters, with the first being the topmost header. The input vector elements can be:

Column names from the split summarised result generated by splitAll()

Settings specified in the settingsColumns argument

group, strata, additional, variable, estimate, and/or settings to refer to all columns within these groups

Any other input to create overall header labels at the specified location.

settingsColumns

A character vector with the names of settings to include in the table.

groupColumn

Columns to use as group labels. By default, the name of the new group will be the tidy* column names separated by ";". To specify a custom group name, use a named list such as: list("newGroupName" = c("variable_name", "variable_level")).

*tidy: The tidy format applied to column names replaces "_" with a space and converts to sentence case. Use renameColumns to customize specific column names.

renameColumns

A named vector to customize column names, e.g., c("Database name" = "cdm_name"). The function renames all column names not specified here into a tidy* format.

type

The type of desired formatted table. Options are: "gt", "flextable", or "tibble".

hide

Columns to drop from the output table. By default, result_id and estimate_type are always dropped.

showMinCellCount

If TRUE, suppressed estimates will be indicated with "<{min_cell_count}", otherwise, the default na defined in .options will be used.

.options

A named list with additional formatting options. visOmopResults::optionsTable() shows allowed arguments and their default values.

split

minCellCount

Value

A tibble, gt, or flextable object.

Examples

result <- mockSummarisedResult()
result |>
  visOmopTable(
    formatEstimateName = c("N%" = "<count> (<percentage>)",
                           "N" = "<count>",
                           "Mean (SD)" = "<mean> (<sd>)"),
    header = c("group"),
    renameColumns = c("Database name" = "cdm_name"),
    groupColumn = strataColumns(result)
  )
#> ! Results have not been suppressed.

  

Database name
      	Variable name
      	Variable level
      	Estimate name
      	
        Cohort name
      
    

cohort1
      	cohort2
    

overall; overall
    

mock
	number subjects
	-
	N
	9,774,459
	3,928,787

	age
	-
	Mean (SD)
	54.50 (3.24)
	13.25 (7.22)

	Medications
	Amoxiciline
	N%
	91,718 (25.36)
	26,115 (58.94)

	
	Ibuprofen
	N%
	11,175 (97.17)
	5,990 (63.95)

<40; Male
    

mock
	number subjects
	-
	N
	8,831,078
	8,108,000

	age
	-
	Mean (SD)
	47.05 (4.96)
	43.34 (6.02)

	Medications
	Amoxiciline
	N%
	2,748 (65.61)
	77,458 (88.68)

	
	Ibuprofen
	N%
	61,946 (27.96)
	16,950 (99.25)

>=40; Male
    

mock
	number subjects
	-
	N
	5,873,406
	3,768,728

	age
	-
	Mean (SD)
	60.19 (9.98)
	19.99 (3.35)

	Medications
	Amoxiciline
	N%
	82,491 (83.93)
	94,141 (30.33)

	
	Ibuprofen
	N%
	69,556 (63.07)
	53,725 (70.90)

<40; Female
    

mock
	number subjects
	-
	N
	4,810,418
	497,417

	age
	-
	Mean (SD)
	76.25 (0.02)
	22.59 (4.11)

	Medications
	Amoxiciline
	N%
	6,248 (83.43)
	86,287 (62.78)

	
	Ibuprofen
	N%
	31,686 (42.33)
	62,064 (75.34)

>=40; Female
    

mock
	number subjects
	-
	N
	2,219,612
	3,873,241

	age
	-
	Mean (SD)
	28.12 (7.32)
	6.63 (7.37)

	Medications
	Amoxiciline
	N%
	90,344 (42.49)
	7,365 (20.50)

	
	Ibuprofen
	N%
	87,919 (32.76)
	35,683 (51.53)

overall; Male
    

mock
	number subjects
	-
	N
	7,843,372
	5,951,709

	age
	-
	Mean (SD)
	54.96 (5.08)
	27.68 (6.58)

	Medications
	Amoxiciline
	N%
	13,248 (28.11)
	88,939 (45.64)

	
	Ibuprofen
	N%
	40,466 (39.38)
	53,646 (61.39)

overall; Female
    

mock
	number subjects
	-
	N
	1,317,417
	2,108,602

	age
	-
	Mean (SD)
	48.29 (0.00)
	38.82 (2.81)

	Medications
	Amoxiciline
	N%
	62,851 (35.79)
	21,432 (63.10)

	
	Ibuprofen
	N%
	63,478 (63.20)
	44,852 (5.37)

<40; overall
    

mock
	number subjects
	-
	N
	8,783,310
	168,998

	age
	-
	Mean (SD)
	90.26 (4.61)
	9.31 (5.30)

	Medications
	Amoxiciline
	N%
	94,496 (94.57)
	33,862 (51.36)

	
	Ibuprofen
	N%
	79,507 (52.94)
	17,454 (44.60)

>=40; overall
    

mock
	number subjects
	-
	N
	8,463,641
	4,623,340

	age
	-
	Mean (SD)
	77.48 (4.83)
	51.71 (4.33)

	Medications
	Amoxiciline
	N%
	12,301 (56.96)
	30,696 (46.28)

	
	Ibuprofen
	N%
	70,879 (84.33)
	65,829 (14.69)