Creates a flextable or gt object from a dataframe

Creates a flextable object from a dataframe using a delimiter to span the header, and allows to easily customise table style.

Usage

formatTable(
  x,
  type = "gt",
  delim = "\n",
  style = "default",
  na = "-",
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  groupColumn = NULL,
  groupAsColumn = FALSE,
  groupOrder = NULL,
  merge = NULL
)

Arguments

x

A dataframe.

type

The desired format of the output table. See tableType() for allowed options. If "tibble", no formatting will be applied.

delim

Delimiter.

style

Named list that specifies how to style the different parts of the gt or flextable table generated. Accepted style entries are: title, subtitle, header, header_name, header_level, column_name, group_label, and body. Alternatively, use "default" to get visOmopResults style, or NULL for gt/flextable style. Keep in mind that styling code is different for gt and flextable. To see the "deafult" style code use tableStyle().

na

How to display missing values. Not used for "datatable".

title

Title of the table, or NULL for no title. Not used for "datatable".

subtitle

Subtitle of the table, or NULL for no subtitle. Not used for "datatable".

caption

Caption for the table, or NULL for no caption. Text in markdown formatting style (e.g. *Your caption here* for caption in italics).

groupColumn

Specifies the columns to use for group labels. By default, the new group name will be a combination of the column names, joined by "_". To assign a custom group name, provide a named list such as: list(newGroupName = c("variable_name", "variable_level"))

groupAsColumn

Whether to display the group labels as a column (TRUE) or rows (FALSE). Not used for "datatable".

groupOrder

Order in which to display group labels. Not used for "datatable".

merge

Names of the columns to merge vertically when consecutive row cells have identical values. Alternatively, use "all_columns" to apply this merging to all columns, or use NULL to indicate no merging. Not used for "datatable".

Value

A flextable object.

A flextable or gt object.

Examples

# Example 1
mockSummarisedResult() |>
  formatEstimateValue(decimals = c(integer = 0, numeric = 1)) |>
  formatHeader(
    header = c("Study strata", "strata_name", "strata_level"),
    includeHeaderName = FALSE
  ) |>
  formatTable(
    type = "flextable",
    style = "default",
    na = "--",
    title = "fxTable example",
    subtitle = NULL,
    caption = NULL,
    groupColumn = "group_level",
    groupAsColumn = TRUE,
    groupOrder = c("cohort1", "cohort2"),
    merge = "all_columns"
  )

fxTable example

group_level

result_id

cdm_name

group_name

variable_name

variable_level

estimate_name

estimate_type

additional_name

additional_level

Study strata

overall

age_group &&& sex

sex

age_group

overall

<40 &&& Male

>=40 &&& Male

<40 &&& Female

>=40 &&& Female

Male

Female

<40

>=40

cohort1

1

mock

cohort_name

number subjects

--

count

integer

overall

overall

7,312,292

3,986,198

4,305,207

1,454,073

4,511,203

5,787,815

704,171

7,423,262

5,492,207

age

--

mean

numeric

overall

overall

77.8

65.8

25.9

91.8

14.9

64.1

20.1

74.4

84.8

sd

numeric

overall

overall

4.4

9.0

7.7

1.3

5.4

1.3

5.1

5.4

2.4

Medications

Amoxiciline

count

integer

overall

overall

36,924

12,667

73,987

95,587

66,867

46,809

53,819

33,095

62,868

percentage

percentage

overall

overall

6.84064193628728

95.8503520581871

24.4483520975336

92.6283334614709

50.2747076796368

93.5517142061144

43.8799632480368

35.0486536743119

47.6504088379443

Ibuprofen

count

integer

overall

overall

13,002

79,191

35,590

96,434

11,868

45,816

52,365

83,706

3,350

percentage

percentage

overall

overall

75.4406645428389

24.2454329272732

62.715786579065

16.8414602987468

44.4789632922038

16.4675234118477

11.8276783032343

36.7464390583336

94.6558905066922

cohort2

1

mock

cohort_name

number subjects

--

count

integer

overall

overall

6,345,443

1,366,286

9,302,335

6,511,186

5,287,093

2,251,154

9,832,669

1,090,488

699,455

age

--

mean

numeric

overall

overall

85.4

80.3

2.4

79.5

88.7

60.0

43.3

92.0

61.6

sd

numeric

overall

overall

8.0

9.5

4.1

5.1

9.8

3.5

4.6

2.2

0.8

Medications

Amoxiciline

count

integer

overall

overall

45,255

88,802

78,420

13,806

28,131

70,695

44,219

67,770

70,071

percentage

percentage

overall

overall

96.6159525094554

29.7161926981062

29.0648308349773

85.9722247580066

19.3312666844577

80.9108371380717

36.5486512426287

25.2699653152376

85.6081013800576

Ibuprofen

count

integer

overall

overall

4,333

79,210

20,270

6,880

86,444

7,709

43,100

97,362

57,096

percentage

percentage

overall

overall

87.5410763313994

26.0058912448585

65.7298437319696

12.1864625485614

31.8848039023578

62.3061030171812

44.6997430175543

70.8293445874006

89.0028999885544

# Example 2
mockSummarisedResult() |>
  formatEstimateValue(decimals = c(integer = 0, numeric = 1)) |>
  formatHeader(header = c("Study strata", "strata_name", "strata_level"),
              includeHeaderName = FALSE) |>
  formatTable(
    type = "gt",
    style = list("header" = list(
      gt::cell_fill(color = "#d9d9d9"),
      gt::cell_text(weight = "bold")),
      "header_level" = list(gt::cell_fill(color = "#e1e1e1"),
                            gt::cell_text(weight = "bold")),
      "column_name" = list(gt::cell_text(weight = "bold")),
      "title" = list(gt::cell_text(weight = "bold"),
                     gt::cell_fill(color = "#c8c8c8")),
      "group_label" = gt::cell_fill(color = "#e1e1e1")),
    na = "--",
    title = "gtTable example",
    subtitle = NULL,
    caption = NULL,
    groupColumn = "group_level",
    groupAsColumn = FALSE,
    groupOrder = c("cohort1", "cohort2"),
    merge = "all_columns"
  )

  

gtTable example
    

      	
        
Study strata

      
    

result_id
      	cdm_name
      	group_name
      	variable_name
      	variable_level
      	estimate_name
      	estimate_type
      	additional_name
      	additional_level
      	
        
overall

      
      	
        
age_group &&& sex

      
      	
        
sex

      
      	
        
age_group

      
    

overall
      	<40 &&& Male
      	>=40 &&& Male
      	<40 &&& Female
      	>=40 &&& Female
      	Male
      	Female
      	<40
      	>=40
    

cohort1
    

1
	mock
	cohort_name
	number subjects
	--
	count
	integer
	overall
	overall
	1,043,065
	6,304,117
	8,415,589
	3,642,624
	6,229,228
	4,607,475
	1,634,123
	5,138,669
	1,432,664

	
	
	age
	--
	mean
	numeric
	overall
	overall
	39.8
	63.2
	94.9
	34.0
	62.7
	82.5
	14.5
	96.7
	8.6

	
	
	
	
	sd
	numeric
	overall
	overall
	9.3
	4.2
	5.0
	3.5
	2.5
	9.6
	3.6
	2.0
	7.3

	
	
	Medications
	Amoxiciline
	count
	integer
	overall
	overall
	92,396
	98,219
	40,527
	26,969
	48,829
	5,106
	248
	49,583
	24,379

	
	
	
	
	percentage
	percentage
	overall
	overall
	56.9740539882332
	69.6488575544208
	70.0767503585666
	3.99406552314758
	43.1522403843701
	81.4231365453452
	56.658627698198
	58.8247172767296
	88.0309032043442

	
	
	
	Ibuprofen
	count
	integer
	overall
	overall
	41,797
	76,393
	79,448
	24,259
	75,917
	2,910
	43,734
	53,041
	42,983

	
	
	
	
	percentage
	percentage
	overall
	overall
	29.3585412902758
	22.4765427643433
	69.0917830914259
	31.2329404754564
	48.3298178529367
	12.179663986899
	69.0007234923542
	31.5731404349208
	81.6187682561576

cohort2
    

1
	mock
	cohort_name
	number subjects
	--
	count
	integer
	overall
	overall
	6,502,987
	1,552,945
	5,083,099
	858,704
	5,030,267
	385,719
	8,240,329
	2,859,856
	2,467,509

	
	
	age
	--
	mean
	numeric
	overall
	overall
	24.0
	63.9
	50.3
	25.1
	97.9
	65.7
	9.0
	63.1
	87.7

	
	
	
	
	sd
	numeric
	overall
	overall
	1.7
	2.3
	8.5
	3.5
	0.7
	0.3
	1.2
	0.1
	7.4

	
	
	Medications
	Amoxiciline
	count
	integer
	overall
	overall
	25,575
	55,232
	25,158
	62,845
	31,612
	92,452
	90,649
	9,197
	73,780

	
	
	
	
	percentage
	percentage
	overall
	overall
	71.9317963346839
	93.6428893357515
	42.943161376752
	28.618732560426
	77.0643183263019
	61.6498636547476
	15.9274747595191
	40.1598005788401
	0.164648867212236

	
	
	
	Ibuprofen
	count
	integer
	overall
	overall
	80,930
	92,256
	67,502
	9,397
	19,746
	47,555
	71,437
	53,808
	87,538

	
	
	
	
	percentage
	percentage
	overall
	overall
	54.5915438327938
	94.7317183716223
	3.8526406744495
	64.764387672767
	7.96190109103918
	43.2930769631639
	76.0893133934587
	29.3758839834481
	10.7331892941147