Generate a formatted table from a data.table

This function combines the functionalities of formatEstimateValue(), formatEstimateName(), formatHeader(), gtTable(), and fxTable() into a single function. While it does not require the input table to be a summarised_result, it does expect specific fields to apply formatting.

Usage

formatTable(
  result,
  formatEstimateName = character(),
  header = character(),
  groupColumn = character(),
  renameColumns = character(),
  type = "gt",
  hide = character(),
  .options = list()
)

Arguments

result

A table to format.

formatEstimateName

A named list of estimate names to join, sorted by computation order. Use <...> to indicate estimate names. This argument requires that the table has estimate_name and estimate_value columns.

header

A vector specifying the elements to include in the header. The order of elements matters, with the first being the topmost header. The vector elements can be column names or labels for overall headers. The table must contain an estimate_value column to pivot the headers.

groupColumn

Columns to use as group labels. By default, the name of the new group will be the tidy* column names separated by ";". To specify a custom group name, use a named list such as: list("newGroupName" = c("variable_name", "variable_level")).

*tidy: The tidy format applied to column names replaces "_" with a space and converts them to sentence case. Use renameColumns to customize specific column names.

renameColumns

A named vector to customize column names, e.g., c("Database name" = "cdm_name"). The function will rename all column names not specified here into a tidy* format.

type

The desired format of the output table. Options are: "gt", "flextable", or "tibble".

hide

Columns to drop from the output table.

.options

A named list with additional formatting options. visOmopResults::optionsTable() shows allowed arguments and their default values.

Value

A tibble, gt, or flextable object.

Examples

result <- mockSummarisedResult()
result |>
  formatTable(
    formatEstimateName = c("N%" = "<count> (<percentage>)",
                           "N" = "<count>",
                           "Mean (SD)" = "<mean> (<sd>)"),
    header = c("Estimate"),
    renameColumns = c("Database name" = "cdm_name"),
    groupColumn = c("strata_name", "strata_level"),
    hide = c("additional_name", "additional_level", "estimate_type", "result_type")
  )

  

Result id
      	Database name
      	Group name
      	Group level
      	Variable name
      	Variable level
      	Estimate name
      	Estimate
    

overall; overall
    

1
	mock
	cohort_name
	cohort1
	number subjects
	-
	N
	6,502,987

	
	
	cohort2
	number subjects
	-
	N
	3,979,936

	
	
	cohort1
	age
	-
	Mean (SD)
	24.05 (1.68)

	
	
	cohort2
	age
	-
	Mean (SD)
	93.03 (9.24)

	
	
	cohort1
	Medications
	Amoxiciline
	N%
	25,575 (71.93)

	
	
	cohort2
	Medications
	Amoxiciline
	N%
	56,974 (41.80)

	
	
	cohort1
	Medications
	Ibuprofen
	N%
	80,930 (54.59)

	
	
	cohort2
	Medications
	Ibuprofen
	N%
	29,359 (7.30)

age_group &&& sex; <40 &&& Male
    

1
	mock
	cohort_name
	cohort1
	number subjects
	-
	N
	1,552,945

	
	
	cohort2
	number subjects
	-
	N
	6,320,929

	
	
	cohort1
	age
	-
	Mean (SD)
	63.89 (2.26)

	
	
	cohort2
	age
	-
	Mean (SD)
	41.95 (9.82)

	
	
	cohort1
	Medications
	Amoxiciline
	N%
	55,232 (93.64)

	
	
	cohort2
	Medications
	Amoxiciline
	N%
	69,649 (76.39)

	
	
	cohort1
	Medications
	Ibuprofen
	N%
	92,256 (94.73)

	
	
	cohort2
	Medications
	Ibuprofen
	N%
	22,477 (35.63)

age_group &&& sex; >=40 &&& Male
    

1
	mock
	cohort_name
	cohort1
	number subjects
	-
	N
	5,083,099

	
	
	cohort2
	number subjects
	-
	N
	9,485,788

	
	
	cohort1
	age
	-
	Mean (SD)
	50.34 (8.46)

	
	
	cohort2
	age
	-
	Mean (SD)
	49.56 (4.05)

	
	
	cohort1
	Medications
	Amoxiciline
	N%
	25,158 (42.94)

	
	
	cohort2
	Medications
	Amoxiciline
	N%
	70,077 (79.45)

	
	
	cohort1
	Medications
	Ibuprofen
	N%
	67,502 (3.85)

	
	
	cohort2
	Medications
	Ibuprofen
	N%
	69,092 (57.16)

age_group &&& sex; <40 &&& Female
    

1
	mock
	cohort_name
	cohort1
	number subjects
	-
	N
	858,704

	
	
	cohort2
	number subjects
	-
	N
	3,401,192

	
	
	cohort1
	age
	-
	Mean (SD)
	25.05 (3.46)

	
	
	cohort2
	age
	-
	Mean (SD)
	35.27 (2.70)

	
	
	cohort1
	Medications
	Amoxiciline
	N%
	62,845 (28.62)

	
	
	cohort2
	Medications
	Amoxiciline
	N%
	3,994 (24.26)

	
	
	cohort1
	Medications
	Ibuprofen
	N%
	9,397 (64.76)

	
	
	cohort2
	Medications
	Ibuprofen
	N%
	31,233 (67.80)

age_group &&& sex; >=40 &&& Female
    

1
	mock
	cohort_name
	cohort1
	number subjects
	-
	N
	5,030,267

	
	
	cohort2
	number subjects
	-
	N
	6,273,660

	
	
	cohort1
	age
	-
	Mean (SD)
	97.90 (0.67)

	
	
	cohort2
	age
	-
	Mean (SD)
	24.87 (4.88)

	
	
	cohort1
	Medications
	Amoxiciline
	N%
	31,612 (77.06)

	
	
	cohort2
	Medications
	Amoxiciline
	N%
	43,152 (75.92)

	
	
	cohort1
	Medications
	Ibuprofen
	N%
	19,746 (7.96)

	
	
	cohort2
	Medications
	Ibuprofen
	N%
	48,330 (69.07)

sex; Male
    

1
	mock
	cohort_name
	cohort1
	number subjects
	-
	N
	385,719

	
	
	cohort2
	number subjects
	-
	N
	8,249,154

	
	
	cohort1
	age
	-
	Mean (SD)
	65.66 (0.27)

	
	
	cohort2
	age
	-
	Mean (SD)
	95.87 (0.51)

	
	
	cohort1
	Medications
	Amoxiciline
	N%
	92,452 (61.65)

	
	
	cohort2
	Medications
	Amoxiciline
	N%
	81,423 (2.91)

	
	
	cohort1
	Medications
	Ibuprofen
	N%
	47,555 (43.29)

	
	
	cohort2
	Medications
	Ibuprofen
	N%
	12,180 (37.34)

sex; Female
    

1
	mock
	cohort_name
	cohort1
	number subjects
	-
	N
	8,240,329

	
	
	cohort2
	number subjects
	-
	N
	1,451,154

	
	
	cohort1
	age
	-
	Mean (SD)
	8.98 (1.18)

	
	
	cohort2
	age
	-
	Mean (SD)
	36.00 (0.02)

	
	
	cohort1
	Medications
	Amoxiciline
	N%
	90,649 (15.93)

	
	
	cohort2
	Medications
	Amoxiciline
	N%
	56,659 (43.73)

	
	
	cohort1
	Medications
	Ibuprofen
	N%
	71,437 (76.09)

	
	
	cohort2
	Medications
	Ibuprofen
	N%
	69,001 (9.80)

age_group; <40
    

1
	mock
	cohort_name
	cohort1
	number subjects
	-
	N
	2,859,856

	
	
	cohort2
	number subjects
	-
	N
	9,671,391

	
	
	cohort1
	age
	-
	Mean (SD)
	63.05 (0.14)

	
	
	cohort2
	age
	-
	Mean (SD)
	19.90 (4.96)

	
	
	cohort1
	Medications
	Amoxiciline
	N%
	9,197 (40.16)

	
	
	cohort2
	Medications
	Amoxiciline
	N%
	58,825 (53.04)

	
	
	cohort1
	Medications
	Ibuprofen
	N%
	53,808 (29.38)

	
	
	cohort2
	Medications
	Ibuprofen
	N%
	31,573 (9.90)

age_group; >=40
    

1
	mock
	cohort_name
	cohort1
	number subjects
	-
	N
	2,467,509

	
	
	cohort2
	number subjects
	-
	N
	860,390

	
	
	cohort1
	age
	-
	Mean (SD)
	87.68 (7.36)

	
	
	cohort2
	age
	-
	Mean (SD)
	72.63 (2.44)

	
	
	cohort1
	Medications
	Amoxiciline
	N%
	73,780 (0.16)

	
	
	cohort2
	Medications
	Amoxiciline
	N%
	88,031 (42.98)

	
	
	cohort1
	Medications
	Ibuprofen
	N%
	87,538 (10.73)

	
	
	cohort2
	Medications
	Ibuprofen
	N%
	81,619 (77.48)