Creates a flextable object from a dataframe

Creates a flextable object from a dataframe using a delimiter to span the header, and allows to easily customise table style.

Usage

fxTable(
  x,
  delim = "\n",
  style = "default",
  na = "-",
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  groupColumn = NULL,
  groupAsColumn = FALSE,
  groupOrder = NULL,
  colsToMergeRows = NULL
)

Arguments

x

A dataframe.

delim

Delimiter.

style

Named list that specifies how to style the different parts of the gt table. Accepted entries are: title, subtitle, header, header_name, header_level, column_name, group_label, and body. Alternatively, use "default" to get visOmopResults style, or NULL for flextable style.

na

How to display missing values.

title

Title of the table, or NULL for no title.

subtitle

Subtitle of the table, or NULL for no subtitle.

caption

Caption for the table, or NULL for no caption. Text in markdown formatting style (e.g. *Your caption here* for caption in italics).

groupColumn

Specifies the columns to use for group labels. By default, the new group name will be a combination of the column names, joined by "_". To assign a custom group name, provide a named list such as: list(newGroupName = c("variable_name", "variable_level"))

groupAsColumn

Whether to display the group labels as a column (TRUE) or rows (FALSE).

groupOrder

Order in which to display group labels.

colsToMergeRows

Names of the columns to merge vertically when consecutive row cells have identical values. Alternatively, use "all_columns" to apply this merging to all columns, or use NULL to indicate no merging.

Value

A flextable object.

A flextable object.

Examples

mockSummarisedResult() |>
  formatEstimateValue(decimals = c(integer = 0, numeric = 1)) |>
  formatHeader(
    header = c("Study strata", "strata_name", "strata_level"),
    includeHeaderName = FALSE
  ) |>
  fxTable(
    style = "default",
    na = "--",
    title = "fxTable example",
    subtitle = NULL,
    caption = NULL,
    groupColumn = "group_level",
    groupAsColumn = TRUE,
    groupOrder = c("cohort1", "cohort2"),
    colsToMergeRows = "all_columns"
  )

fxTable example

group_level

result_id

cdm_name

group_name

variable_name

variable_level

estimate_name

estimate_type

additional_name

additional_level

Study strata

overall

age_group &&& sex

sex

age_group

overall

<40 &&& Male

>=40 &&& Male

<40 &&& Female

>=40 &&& Female

Male

Female

<40

>=40

cohort1

1

mock

cohort_name

number subjects

--

count

integer

overall

overall

185,314

8,509,742

5,014,234

1,661,618

3,610,061

6,520,002

8,741,768

7,283,057

4,124,729

age

--

mean

numeric

overall

overall

17.8

36.3

39.1

90.4

27.6

13.9

45.1

11.0

88.1

sd

numeric

overall

overall

0.2

9.3

0.8

0.3

3.1

7.8

2.8

1.7

4.3

Medications

Amoxiciline

count

integer

overall

overall

77,303

2,212

81,502

30,383

93,180

54,043

25,482

47,759

70,446

percentage

percentage

overall

overall

39.5447959192097

17.4868220463395

32.7313161687925

61.4753181813285

7.91664109565318

51.9060970284045

30.6281309342012

67.1395304379985

39.3671740312129

Ibuprofen

count

integer

overall

overall

23,838

41,168

74,607

72,345

61,452

47,398

57,846

24,159

11,438

percentage

percentage

overall

overall

87.3739800183102

88.2156537845731

10.7689367607236

57.4544371105731

99.6033379342407

85.0553432479501

43.7655055196956

32.6429924694821

33.596885856241

cohort2

1

mock

cohort_name

number subjects

--

count

integer

overall

overall

1,592,293

8,457,036

8,843,691

3,959,981

381,408

9,513,960

7,837,275

2,194,397

3,309,320

age

--

mean

numeric

overall

overall

53.4

73.9

72.0

26.4

67.4

21.0

69.0

19.5

78.2

sd

numeric

overall

overall

8.1

6.9

7.8

0.9

7.4

4.0

3.0

7.4

4.7

Medications

Amoxiciline

count

integer

overall

overall

751

77,241

26,579

97,104

19,053

1,970

61,238

52,573

63,452

percentage

percentage

overall

overall

74.7848246712238

13.8926521642134

37.7343045547605

64.1383435344324

99.4194674305618

36.8559058057144

33.7348550790921

91.3685443112627

90.9420361043885

Ibuprofen

count

integer

overall

overall

29,834

48,412

56,741

75,395

54,543

75,968

21,297

83,927

40,670

percentage

percentage

overall

overall

23.927381564863

42.137112445198

51.2136830715463

18.5446366202086

32.6400522142649

81.7073642974719

74.8737998073921

53.1741869403049

96.1571368388832