Creats a flextable object from a dataframe

Creats a flextable object from a dataframe using as delimiter (delim) to span the header, and the specified styles for different parts of the table.

Usage

fxTable(
  x,
  delim = "\n",
  style = "default",
  na = "-",
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  groupNameCol = NULL,
  groupNameAsColumn = FALSE,
  groupOrder = NULL
)

Arguments

x

A dataframe.

delim

Delimiter.

style

Named list that specifies how to style the different parts of the gt table. Accepted entries are: title, subtitle, header, header_name, header_level, column_name, group_label, and body. Alternatively, use "default" to get visOmopResult style, or NULL for gt style

na

How to display missing values.

title

Title of the table, or NULL for no title.

subtitle

Subtitle of the table, or NULL for no subtitle.

caption

Caption for the table, or NULL for no caption. Text in markdown formatting style (e.g. *Your caption here* for caption in italics).

groupNameCol

Column to use as group labels.

groupNameAsColumn

Whether to display the group labels as a column (TRUE) or rows (FALSE).

groupOrder

Order in which to display group labels.

Value

flextable object

A flextable.

Examples

# \donttest{
mockSummarisedResult() |>
  formatEstimateValue(decimals = c(integer = 0, numeric = 1)) |>
  formatTable(header = c("Study strata", "strata_level"),
              includeHeaderName = FALSE) |>
  fxTable(
    style = "default",
    na = "--",
    title = "fxTable example",
    subtitle = NULL,
    caption = NULL,
    groupNameCol = "group_level",
    groupNameAsColumn = TRUE,
    groupOrder = c("cohort1", "cohort2")
 )

fxTable example

group_level

cdm_name

result_type

package_name

package_version

group_name

strata_name

variable_name

variable_level

estimate_name

estimate_type

additional_name

additional_level

Study strata

overall

<40 and Male

>=40 and Male

<40 and Female

>=40 and Female

Male

Female

<40

>=40

cohort1

mock

--

visOmopResults

0.0.1

cohort_name

overall

number subjects

--

count

integer

overall

overall

1,163,092

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

number subjects

--

count

integer

overall

overall

--

7,154,231

2,261,087

1,333,485

9,272,818

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

number subjects

--

count

integer

overall

overall

--

--

--

--

--

8,941,227

2,036,644

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

number subjects

--

count

integer

overall

overall

--

--

--

--

--

--

--

2,569,829

6,140,056

mock

--

visOmopResults

0.0.1

cohort_name

overall

age

--

mean

numeric

overall

overall

7.6

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

age

--

mean

numeric

overall

overall

--

20.6

49.6

26.2

41.6

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

age

--

mean

numeric

overall

overall

--

--

--

--

--

43.9

18.2

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

age

--

mean

numeric

overall

overall

--

--

--

--

--

--

--

7.8

1.8

mock

--

visOmopResults

0.0.1

cohort_name

overall

age

--

sd

numeric

overall

overall

5.2

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

age

--

sd

numeric

overall

overall

--

5.0

0.1

9.6

4.5

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

age

--

sd

numeric

overall

overall

--

--

--

--

--

9.8

5.4

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

age

--

sd

numeric

overall

overall

--

--

--

--

--

--

--

3.6

5.0

mock

--

visOmopResults

0.0.1

cohort_name

overall

Medications

Amoxiciline

count

integer

overall

overall

37,896

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

Medications

Amoxiciline

count

integer

overall

overall

--

35,953

84,423

45,675

17,903

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

Medications

Amoxiciline

count

integer

overall

overall

--

--

--

--

--

96,621

71,341

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

Medications

Amoxiciline

count

integer

overall

overall

--

--

--

--

--

--

--

2,489

62,646

mock

--

visOmopResults

0.0.1

cohort_name

overall

Medications

Amoxiciline

percentage

percentage

overall

overall

92.1526176389307

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

Medications

Amoxiciline

percentage

percentage

overall

overall

--

67.5661901943386

35.0588162429631

0.865807477384806

43.0898792576045

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

Medications

Amoxiciline

percentage

percentage

overall

overall

--

--

--

--

--

72.493509715423

41.2256387062371

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

Medications

Amoxiciline

percentage

percentage

overall

overall

--

--

--

--

--

--

--

55.0862395204604

15.6360570108518

mock

--

visOmopResults

0.0.1

cohort_name

overall

Medications

Ibuprofen

count

integer

overall

overall

80,970

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

Medications

Ibuprofen

count

integer

overall

overall

--

55,938

73,373

34,970

98,267

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

Medications

Ibuprofen

count

integer

overall

overall

--

--

--

--

--

40,885

36,087

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

Medications

Ibuprofen

count

integer

overall

overall

--

--

--

--

--

--

--

65,251

14,356

mock

--

visOmopResults

0.0.1

cohort_name

overall

Medications

Ibuprofen

percentage

percentage

overall

overall

73.110364517197

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

Medications

Ibuprofen

percentage

percentage

overall

overall

--

85.2432376472279

97.5818040780723

11.3047500839457

97.0262279734015

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

Medications

Ibuprofen

percentage

percentage

overall

overall

--

--

--

--

--

64.8082781815901

5.6073043262586

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

Medications

Ibuprofen

percentage

percentage

overall

overall

--

--

--

--

--

--

--

47.3137095803395

29.4625013135374

cohort2

mock

--

visOmopResults

0.0.1

cohort_name

overall

number subjects

--

count

integer

overall

overall

4,414,992

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

number subjects

--

count

integer

overall

overall

--

3,155,947

1,013,138

2,726,462

6,536,994

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

number subjects

--

count

integer

overall

overall

--

--

--

--

--

9,278,701

266,247

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

number subjects

--

count

integer

overall

overall

--

--

--

--

--

--

--

5,594,912

8,541,737

mock

--

visOmopResults

0.0.1

cohort_name

overall

age

--

mean

numeric

overall

overall

72.8

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

age

--

mean

numeric

overall

overall

--

97.1

53.5

16.6

5.8

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

age

--

mean

numeric

overall

overall

--

--

--

--

--

45.9

62.3

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

age

--

mean

numeric

overall

overall

--

--

--

--

--

--

--

59.5

74.3

mock

--

visOmopResults

0.0.1

cohort_name

overall

age

--

sd

numeric

overall

overall

0.7

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

age

--

sd

numeric

overall

overall

--

9.7

6.5

9.8

5.4

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

age

--

sd

numeric

overall

overall

--

--

--

--

--

7.6

5.6

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

age

--

sd

numeric

overall

overall

--

--

--

--

--

--

--

7.8

6.4

mock

--

visOmopResults

0.0.1

cohort_name

overall

Medications

Amoxiciline

count

integer

overall

overall

82,792

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

Medications

Amoxiciline

count

integer

overall

overall

--

64,307

70,848

87,073

28,181

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

Medications

Amoxiciline

count

integer

overall

overall

--

--

--

--

--

17,322

10,967

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

Medications

Amoxiciline

count

integer

overall

overall

--

--

--

--

--

--

--

58,622

85,072

mock

--

visOmopResults

0.0.1

cohort_name

overall

Medications

Amoxiciline

percentage

percentage

overall

overall

4.30012932047248

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

Medications

Amoxiciline

percentage

percentage

overall

overall

--

68.464411306195

51.4402497094125

94.568784837611

84.309694217518

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

Medications

Amoxiciline

percentage

percentage

overall

overall

--

--

--

--

--

6.48830933496356

76.2110914569348

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

Medications

Amoxiciline

percentage

percentage

overall

overall

--

--

--

--

--

--

--

7.83842522650957

13.3507662685588

mock

--

visOmopResults

0.0.1

cohort_name

overall

Medications

Ibuprofen

count

integer

overall

overall

68,080

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

Medications

Ibuprofen

count

integer

overall

overall

--

85,929

83,863

88,131

23,770

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

Medications

Ibuprofen

count

integer

overall

overall

--

--

--

--

--

57,863

89,790

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

Medications

Ibuprofen

count

integer

overall

overall

--

--

--

--

--

--

--

34,443

76,086

mock

--

visOmopResults

0.0.1

cohort_name

overall

Medications

Ibuprofen

percentage

percentage

overall

overall

61.0299621941522

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

Medications

Ibuprofen

percentage

percentage

overall

overall

--

12.1108880266547

62.9416507901624

71.1955302860588

61.2082786625251

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

Medications

Ibuprofen

percentage

percentage

overall

overall

--

--

--

--

--

3.44292330555618

66.5890787029639

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

Medications

Ibuprofen

percentage

percentage

overall

overall

--

--

--

--

--

--

--

65.2967276517302

45.1858439715579

# }