Creats a flextable object from a dataframe

Creats a flextable object from a dataframe using as delimiter (delim) to span the header, and the specified styles for different parts of the table.

Usage

fxTable(
  x,
  delim = "\n",
  style = list(header = list(cell = officer::fp_cell(background.color = "#c8c8c8"), text
    = officer::fp_text(bold = TRUE)), header_name = list(cell =
    officer::fp_cell(background.color = "#d9d9d9"), text = officer::fp_text(bold =
    TRUE)), header_level = list(cell = officer::fp_cell(background.color = "#e1e1e1"),
    text = officer::fp_text(bold = TRUE)), column_name = list(text =
    officer::fp_text(bold = TRUE)), group_label = list(cell =
    officer::fp_cell(background.color = "#e9ecef"), text = officer::fp_text(bold =
    TRUE))),
  na = "-",
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  groupNameCol = NULL,
  groupNameAsColumn = FALSE,
  groupOrder = NULL
)

Arguments

x

A dataframe.

delim

Delimiter.

style

Named list that specifies how to style the different parts of the flextable. Accepted entries are: title, subtitle, header, header_name, header_level, column_name, group_label, and body.

na

How to display missing values.

title

Title of the table, or NULL for no title.

subtitle

Subtitle of the table, or NULL for no subtitle.

caption

Caption for the table, or NULL for no caption. Text in markdown formatting style (e.g. *Your caption here* for caption in italics).

groupNameCol

Column to use as group labels.

groupNameAsColumn

Whether to display the group labels as a column (TRUE) or rows (FALSE).

groupOrder

Order in which to display group labels.

Value

flextable object

A flextable.

Examples

# \donttest{
mockSummarisedResult() |>
  formatEstimateValue(decimals = c(integer = 0, numeric = 1)) |>
  formatTable(header = c("Study strata", "strata_level"),
              includeHeaderName = FALSE) |>
  fxTable(
    style = list(
      "header" = list(
        "cell" = officer::fp_cell(background.color = "#c8c8c8"),
        "text" = officer::fp_text(bold = TRUE)),
      "header_level" = list(
        "cell" = officer::fp_cell(background.color = "#e1e1e1"),
        "text" = officer::fp_text(bold = TRUE)),
      "group_label" = list(
        "cell" = officer::fp_cell(
          background.color = "#e9ecef",
          border = officer::fp_border(width = 1, color = "gray")),
        "text" = officer::fp_text(bold = TRUE))
    ),
    na = "--",
    title = "fxTable example",
    subtitle = NULL,
    caption = NULL,
    groupNameCol = "group_level",
    groupNameAsColumn = TRUE,
    groupOrder = c("cohort1", "cohort2")
 )

fxTable example

group_level

cdm_name

result_type

package_name

package_version

group_name

strata_name

variable_name

variable_level

estimate_name

estimate_type

additional_name

additional_level

Study strata

overall

<40 and Male

>=40 and Male

<40 and Female

>=40 and Female

Male

Female

<40

>=40

cohort1

mock

--

visOmopResults

0.0.1

cohort_name

overall

number subjects

--

count

integer

overall

overall

6,752,085

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

number subjects

--

count

integer

overall

overall

--

9,462,948

1,962,195

9,686,375

3,870,963

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

number subjects

--

count

integer

overall

overall

--

--

--

--

--

6,503,439

8,145,962

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

number subjects

--

count

integer

overall

overall

--

--

--

--

--

--

--

  709,648

5,268,303

mock

--

visOmopResults

0.0.1

cohort_name

overall

age

--

mean

numeric

overall

overall

83.6

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

age

--

mean

numeric

overall

overall

--

60.5

90.7

 3.6

13.1

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

age

--

mean

numeric

overall

overall

--

--

--

--

--

 9.4

69.7

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

age

--

mean

numeric

overall

overall

--

--

--

--

--

--

--

40.6

 6.6

mock

--

visOmopResults

0.0.1

cohort_name

overall

age

--

sd

numeric

overall

overall

 2.9

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

age

--

sd

numeric

overall

overall

--

 0.1

 8.5

 4.1

 7.5

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

age

--

sd

numeric

overall

overall

--

--

--

--

--

 7.2

 5.2

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

age

--

sd

numeric

overall

overall

--

--

--

--

--

--

--

 9.7

 3.6

mock

--

visOmopResults

0.0.1

cohort_name

overall

Medications

Amoxiciline

count

integer

overall

overall

   37,864

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

Medications

Amoxiciline

count

integer

overall

overall

--

   17,439

   60,375

   80,546

    3,668

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

Medications

Amoxiciline

count

integer

overall

overall

--

--

--

--

--

   73,291

   21,527

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

Medications

Amoxiciline

count

integer

overall

overall

--

--

--

--

--

--

--

    1,600

   12,860

mock

--

visOmopResults

0.0.1

cohort_name

overall

Medications

Amoxiciline

percentage

percentage

overall

overall

16.5010989643633

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

Medications

Amoxiciline

percentage

percentage

overall

overall

--

56.6909784451127

89.8077371064574

59.4472281169146

83.168991911225

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

Medications

Amoxiciline

percentage

percentage

overall

overall

--

--

--

--

--

59.340836503543

77.8970633633435

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

Medications

Amoxiciline

percentage

percentage

overall

overall

--

--

--

--

--

--

--

39.7771646967158

84.9882774055004

cohort2

mock

--

visOmopResults

0.0.1

cohort_name

overall

number subjects

--

count

integer

overall

overall

7,634,748

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

number subjects

--

count

integer

overall

overall

--

4,353,866

5,524,723

2,040,306

  310,260

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

number subjects

--

count

integer

overall

overall

--

--

--

--

--

9,697,071

1,786,131

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

number subjects

--

count

integer

overall

overall

--

--

--

--

--

--

--

7,782,928

8,857,108

mock

--

visOmopResults

0.0.1

cohort_name

overall

age

--

mean

numeric

overall

overall

12.6

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

age

--

mean

numeric

overall

overall

--

93.7

21.6

66.6

20.4

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

age

--

mean

numeric

overall

overall

--

--

--

--

--

54.9

82.8

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

age

--

mean

numeric

overall

overall

--

--

--

--

--

--

--

12.8

26.1

mock

--

visOmopResults

0.0.1

cohort_name

overall

age

--

sd

numeric

overall

overall

 1.7

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

age

--

sd

numeric

overall

overall

--

 7.8

 0.3

 7.9

 1.7

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

age

--

sd

numeric

overall

overall

--

--

--

--

--

 0.3

 7.9

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

age

--

sd

numeric

overall

overall

--

--

--

--

--

--

--

 8.3

 9.7

mock

--

visOmopResults

0.0.1

cohort_name

overall

Medications

Amoxiciline

count

integer

overall

overall

   68,606

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

Medications

Amoxiciline

count

integer

overall

overall

--

   64,194

   32,734

   38,748

   70,750

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

Medications

Amoxiciline

count

integer

overall

overall

--

--

--

--

--

   69,766

   92,566

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

Medications

Amoxiciline

count

integer

overall

overall

--

--

--

--

--

--

--

   45,902

   59,595

mock

--

visOmopResults

0.0.1

cohort_name

overall

Medications

Amoxiciline

percentage

percentage

overall

overall

74.1845581447706

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

Medications

Amoxiciline

percentage

percentage

overall

overall

--

31.7790219327435

11.1680223839357

10.1095365360379

80.0097748171538

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

Medications

Amoxiciline

percentage

percentage

overall

overall

--

--

--

--

--

37.9946742206812

5.27009717188776

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

Medications

Amoxiciline

percentage

percentage

overall

overall

--

--

--

--

--

--

--

98.6508697737008

60.4173966916278

# }