Creates a flextable object from a dataframe

Creates a flextable object from a dataframe using a delimiter to span the header, and allows to easily customise table style.

Usage

fxTable(
  x,
  delim = "\n",
  style = "default",
  na = "-",
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  groupNameCol = NULL,
  groupNameAsColumn = FALSE,
  groupOrder = NULL,
  colsToMergeRows = NULL
)

Arguments

x

A dataframe.

delim

Delimiter.

style

Named list that specifies how to style the different parts of the gt table. Accepted entries are: title, subtitle, header, header_name, header_level, column_name, group_label, and body. Alternatively, use "default" to get visOmopResults style, or NULL for flextable style.

na

How to display missing values.

title

Title of the table, or NULL for no title.

subtitle

Subtitle of the table, or NULL for no subtitle.

caption

Caption for the table, or NULL for no caption. Text in markdown formatting style (e.g. *Your caption here* for caption in italics).

groupNameCol

Column to use as group labels.

groupNameAsColumn

Whether to display the group labels as a column (TRUE) or rows (FALSE).

groupOrder

Order in which to display group labels.

colsToMergeRows

Names of the columns to merge vertically when consecutive row cells have identical values. Alternatively, use "all_columns" to apply this merging to all columns, or use NULL to indicate no merging.

Value

A flextable object.

A flextable object.

Examples

mockSummarisedResult() |>
  formatEstimateValue(decimals = c(integer = 0, numeric = 1)) |>
  formatHeader(header = c("Study strata", "strata_name", "strata_level"),
               includeHeaderName = FALSE) |>
  fxTable(
    style = "default",
    na = "--",
    title = "fxTable example",
    subtitle = NULL,
    caption = NULL,
    groupNameCol = "group_level",
    groupNameAsColumn = TRUE,
    groupOrder = c("cohort1", "cohort2"),
    colsToMergeRows = "all_columns"
 )

fxTable example

group_level

result_id

cdm_name

result_type

package_name

package_version

group_name

variable_name

variable_level

estimate_name

estimate_type

additional_name

additional_level

Study strata

overall

age_group &&& sex

sex

age_group

overall

<40 &&& Male

>=40 &&& Male

<40 &&& Female

>=40 &&& Female

Male

Female

<40

>=40

cohort1

1

mock

mock_summarised_result

visOmopResults

0.2.0

cohort_name

number subjects

--

count

integer

overall

overall

5,492,207

6,345,443

1,366,286

9,302,335

6,511,186

5,287,093

2,251,154

9,832,669

1,090,488

age

--

mean

numeric

overall

overall

84.8

85.4

80.3

2.4

79.5

88.7

60.0

43.3

92.0

sd

numeric

overall

overall

2.4

8.0

9.5

4.1

5.1

9.8

3.5

4.6

2.2

Medications

Amoxiciline

count

integer

overall

overall

62,868

45,255

88,802

78,420

13,806

28,131

70,695

44,219

67,770

percentage

percentage

overall

overall

47.6504088379443

96.6159525094554

29.7161926981062

29.0648308349773

85.9722247580066

19.3312666844577

80.9108371380717

36.5486512426287

25.2699653152376

Ibuprofen

count

integer

overall

overall

3,350

4,333

79,210

20,270

6,880

86,444

7,709

43,100

97,362

percentage

percentage

overall

overall

94.6558905066922

87.5410763313994

26.0058912448585

65.7298437319696

12.1864625485614

31.8848039023578

62.3061030171812

44.6997430175543

70.8293445874006

cohort2

1

mock

mock_summarised_result

visOmopResults

0.2.0

cohort_name

number subjects

--

count

integer

overall

overall

699,455

7,776,455

6,583,222

2,592,027

9,183,378

1,487,103

6,406,805

2,012,384

7,436,213

age

--

mean

numeric

overall

overall

61.6

43.6

89.7

76.9

12.8

53.9

12.9

51.1

54.4

sd

numeric

overall

overall

0.8

3.7

1.3

7.4

9.6

6.7

4.7

5.4

3.3

Medications

Amoxiciline

count

integer

overall

overall

70,071

6,841

95,850

24,448

92,628

50,275

93,552

43,880

35,049

percentage

percentage

overall

overall

85.6081013800576

13.0024369573221

79.1907660430297

35.5901510221884

96.4343837695196

11.8683209177107

45.8160649519414

52.3650436662138

83.7061725789681

Ibuprofen

count

integer

overall

overall

57,096

75,441

24,245

62,716

16,841

44,479

16,468

11,828

36,746

percentage

percentage

overall

overall

89.0028999885544

10.4306503897533

63.0411718972027

84.1558910673484

36.4262431394309

62.2922778129578

46.0747514152899

16.3412279449403

51.3866862049326