Creates a flextable object from a dataframe

Creates a flextable object from a dataframe using a delimiter to span the header, and allows to easily customise table style.

Usage

fxTable(
  x,
  delim = "\n",
  style = "default",
  na = "-",
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  groupNameCol = NULL,
  groupNameAsColumn = FALSE,
  groupOrder = NULL,
  colsToMergeRows = NULL
)

Arguments

x

A dataframe.

delim

Delimiter.

style

Named list that specifies how to style the different parts of the gt table. Accepted entries are: title, subtitle, header, header_name, header_level, column_name, group_label, and body. Alternatively, use "default" to get visOmopResults style, or NULL for flextable style.

na

How to display missing values.

title

Title of the table, or NULL for no title.

subtitle

Subtitle of the table, or NULL for no subtitle.

caption

Caption for the table, or NULL for no caption. Text in markdown formatting style (e.g. *Your caption here* for caption in italics).

groupNameCol

Column to use as group labels.

groupNameAsColumn

Whether to display the group labels as a column (TRUE) or rows (FALSE).

groupOrder

Order in which to display group labels.

colsToMergeRows

Names of the columns to merge vertically when consecutive row cells have identical values. Alternatively, use "all_columns" to apply this merging to all columns, or use NULL to indicate no merging.

Value

A flextable object.

A flextable object.

Examples

mockSummarisedResult() |>
  formatEstimateValue(decimals = c(integer = 0, numeric = 1)) |>
  formatHeader(header = c("Study strata", "strata_name", "strata_level"),
               includeHeaderName = FALSE) |>
  fxTable(
    style = "default",
    na = "--",
    title = "fxTable example",
    subtitle = NULL,
    caption = NULL,
    groupNameCol = "group_level",
    groupNameAsColumn = TRUE,
    groupOrder = c("cohort1", "cohort2"),
    colsToMergeRows = "all_columns"
 )

fxTable example

group_level

result_id

cdm_name

result_type

package_name

package_version

group_name

variable_name

variable_level

estimate_name

estimate_type

additional_name

additional_level

Study strata

overall

age_group &&& sex

sex

age_group

overall

<40 &&& Male

>=40 &&& Male

<40 &&& Female

>=40 &&& Female

Male

Female

<40

>=40

cohort1

1

mock

mock_summarised_result

visOmopResults

0.1.0

cohort_name

number subjects

--

count

integer

overall

overall

5,168,041

6,766,108

8,809,331

7,312,292

3,986,198

4,305,207

1,454,073

4,511,203

5,787,815

age

--

mean

numeric

overall

overall

98.3

10.9

7.0

77.8

65.8

25.9

91.8

14.9

64.1

sd

numeric

overall

overall

4.3

9.2

6.2

4.4

9.0

7.7

1.3

5.4

1.3

Medications

Amoxiciline

count

integer

overall

overall

45,584

22,002

8,183

36,924

12,667

73,987

95,587

66,867

46,809

percentage

percentage

overall

overall

44.2186920205131

67.7698523038998

70.0714985607192

6.84064193628728

95.8503520581871

24.4483520975336

92.6283334614709

50.2747076796368

93.5517142061144

Ibuprofen

count

integer

overall

overall

36,549

25,270

85,608

13,002

79,191

35,590

96,434

11,868

45,816

percentage

percentage

overall

overall

43.100341851823

97.3622557939962

57.0957343792543

75.4406645428389

24.2454329272732

62.715786579065

16.8414602987468

44.4789632922038

16.4675234118477

cohort2

1

mock

mock_summarised_result

visOmopResults

0.1.0

cohort_name

number subjects

--

count

integer

overall

overall

704,171

7,423,262

5,492,207

6,345,443

1,366,286

9,302,335

6,511,186

5,287,093

2,251,154

age

--

mean

numeric

overall

overall

20.1

74.4

84.8

85.4

80.3

2.4

79.5

88.7

60.0

sd

numeric

overall

overall

5.1

5.4

2.4

8.0

9.5

4.1

5.1

9.8

3.5

Medications

Amoxiciline

count

integer

overall

overall

53,819

33,095

62,868

45,255

88,802

78,420

13,806

28,131

70,695

percentage

percentage

overall

overall

43.8799632480368

35.0486536743119

47.6504088379443

96.6159525094554

29.7161926981062

29.0648308349773

85.9722247580066

19.3312666844577

80.9108371380717

Ibuprofen

count

integer

overall

overall

52,365

83,706

3,350

4,333

79,210

20,270

6,880

86,444

7,709

percentage

percentage

overall

overall

11.8276783032343

36.7464390583336

94.6558905066922

87.5410763313994

26.0058912448585

65.7298437319696

12.1864625485614

31.8848039023578

62.3061030171812