Creats a flextable object from a dataframe

Creats a flextable object from a dataframe using as delimiter (delim) to span the header, and the specified styles for different parts of the table.

Usage

fxTable(
  x,
  delim = "\n",
  style = "default",
  na = "-",
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  groupNameCol = NULL,
  groupNameAsColumn = FALSE,
  groupOrder = NULL
)

Arguments

x

A dataframe.

delim

Delimiter.

style

Named list that specifies how to style the different parts of the gt table. Accepted entries are: title, subtitle, header, header_name, header_level, column_name, group_label, and body. Alternatively, use "default" to get visOmopResult style, or NULL for gt style

na

How to display missing values.

title

Title of the table, or NULL for no title.

subtitle

Subtitle of the table, or NULL for no subtitle.

caption

Caption for the table, or NULL for no caption. Text in markdown formatting style (e.g. *Your caption here* for caption in italics).

groupNameCol

Column to use as group labels.

groupNameAsColumn

Whether to display the group labels as a column (TRUE) or rows (FALSE).

groupOrder

Order in which to display group labels.

Value

flextable object

A flextable.

Examples

# \donttest{
mockSummarisedResult() |>
  formatEstimateValue(decimals = c(integer = 0, numeric = 1)) |>
  formatTable(header = c("Study strata", "strata_level"),
              includeHeaderName = FALSE) |>
  fxTable(
    style = "default",
    na = "--",
    title = "fxTable example",
    subtitle = NULL,
    caption = NULL,
    groupNameCol = "group_level",
    groupNameAsColumn = TRUE,
    groupOrder = c("cohort1", "cohort2")
 )

fxTable example

group_level

cdm_name

result_type

package_name

package_version

group_name

strata_name

variable_name

variable_level

estimate_name

estimate_type

additional_name

additional_level

Study strata

overall

<40 and Male

>=40 and Male

<40 and Female

>=40 and Female

Male

Female

<40

>=40

cohort1

mock

--

visOmopResults

0.0.1

cohort_name

overall

number subjects

--

count

integer

overall

overall

1,487,827

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

number subjects

--

count

integer

overall

overall

--

5,387,602

1,257,892

9,631,818

  469,350

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

number subjects

--

count

integer

overall

overall

--

--

--

--

--

1,614,092

9,364,874

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

number subjects

--

count

integer

overall

overall

--

--

--

--

--

--

--

9,706,814

7,102,741

mock

--

visOmopResults

0.0.1

cohort_name

overall

age

--

mean

numeric

overall

overall

87.3

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

age

--

mean

numeric

overall

overall

--

51.6

86.9

85.7

34.6

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

age

--

mean

numeric

overall

overall

--

--

--

--

--

 0.0

20.5

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

age

--

mean

numeric

overall

overall

--

--

--

--

--

--

--

94.5

28.1

mock

--

visOmopResults

0.0.1

cohort_name

overall

age

--

sd

numeric

overall

overall

 8.9

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

age

--

sd

numeric

overall

overall

--

 8.4

 7.3

 0.4

 3.3

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

age

--

sd

numeric

overall

overall

--

--

--

--

--

 7.5

 9.1

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

age

--

sd

numeric

overall

overall

--

--

--

--

--

--

--

 2.0

 4.7

mock

--

visOmopResults

0.0.1

cohort_name

overall

Medications

Amoxiciline

count

integer

overall

overall

   26,331

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

Medications

Amoxiciline

count

integer

overall

overall

--

   49,785

    1,953

    8,718

   99,596

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

Medications

Amoxiciline

count

integer

overall

overall

--

--

--

--

--

   94,213

   12,437

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

Medications

Amoxiciline

count

integer

overall

overall

--

--

--

--

--

--

--

   11,519

   36,669

mock

--

visOmopResults

0.0.1

cohort_name

overall

Medications

Amoxiciline

percentage

percentage

overall

overall

71.175341703929

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

Medications

Amoxiciline

percentage

percentage

overall

overall

--

96.8462549848482

66.1077789729461

79.6250590356067

7.32091320678592

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

Medications

Amoxiciline

percentage

percentage

overall

overall

--

--

--

--

--

88.4742884431034

82.6700376812369

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

Medications

Amoxiciline

percentage

percentage

overall

overall

--

--

--

--

--

--

--

72.5763959344476

36.7329021682963

cohort2

mock

--

visOmopResults

0.0.1

cohort_name

overall

number subjects

--

count

integer

overall

overall

8,850,909

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

number subjects

--

count

integer

overall

overall

--

9,768,753

  348,996

4,387,573

6,444,698

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

number subjects

--

count

integer

overall

overall

--

--

--

--

--

9,948,324

3,146,626

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

number subjects

--

count

integer

overall

overall

--

--

--

--

--

--

--

8,558,820

5,403,377

mock

--

visOmopResults

0.0.1

cohort_name

overall

age

--

mean

numeric

overall

overall

88.1

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

age

--

mean

numeric

overall

overall

--

96.1

 0.4

55.3

21.9

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

age

--

mean

numeric

overall

overall

--

--

--

--

--

64.7

72.8

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

age

--

mean

numeric

overall

overall

--

--

--

--

--

--

--

87.3

38.2

mock

--

visOmopResults

0.0.1

cohort_name

overall

age

--

sd

numeric

overall

overall

 3.9

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

age

--

sd

numeric

overall

overall

--

 3.9

 2.2

 0.1

 3.6

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

age

--

sd

numeric

overall

overall

--

--

--

--

--

 5.3

 8.4

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

age

--

sd

numeric

overall

overall

--

--

--

--

--

--

--

 1.4

 3.8

mock

--

visOmopResults

0.0.1

cohort_name

overall

Medications

Amoxiciline

count

integer

overall

overall

   47,104

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

Medications

Amoxiciline

count

integer

overall

overall

--

   78,417

   49,039

   10,195

   94,346

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

Medications

Amoxiciline

count

integer

overall

overall

--

--

--

--

--

   70,614

   50,226

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

Medications

Amoxiciline

count

integer

overall

overall

--

--

--

--

--

--

--

   52,993

   25,843

mock

--

visOmopResults

0.0.1

cohort_name

overall

Medications

Amoxiciline

percentage

percentage

overall

overall

52.3428416578099

--

--

--

--

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group and sex

Medications

Amoxiciline

percentage

percentage

overall

overall

--

46.9878912670538

9.85663856845349

51.6034910222515

44.9733827495947

--

--

--

--

mock

--

visOmopResults

0.0.1

cohort_name

sex

Medications

Amoxiciline

percentage

percentage

overall

overall

--

--

--

--

--

54.946154775098

64.9506189860404

--

--

mock

--

visOmopResults

0.0.1

cohort_name

age_group

Medications

Amoxiciline

percentage

percentage

overall

overall

--

--

--

--

--

--

--

15.9034826094285

29.3750792043284

# }