Creates a gt object from a dataframe

Creates a flextable object from a dataframe using a delimiter to span the header, and allows to easily customise table style.

Usage

gtTable(
  x,
  delim = "\n",
  style = "default",
  na = "-",
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  groupColumn = NULL,
  groupNameCol = lifecycle::deprecated(),
  groupAsColumn = FALSE,
  groupNameAsColumn = lifecycle::deprecated(),
  groupOrder = NULL,
  colsToMergeRows = NULL
)

Arguments

x

A dataframe.

delim

Delimiter.

style

Named list that specifies how to style the different parts of the gt table. Accepted entries are: title, subtitle, header, header_name, header_level, column_name, group_label, and body. Alternatively, use "default" to get visOmopResults style, or NULL for gt style

na

How to display missing values.

title

Title of the table, or NULL for no title.

subtitle

Subtitle of the table, or NULL for no subtitle.

caption

Caption for the table, or NULL for no caption. Text in markdown formatting style (e.g. *Your caption here* for caption in italics).

groupColumn

Specifies the columns to use for group labels. By default, the new group name will be a combination of the column names, joined by "_". To assign a custom group name, provide a named list such as: list(newGroupName = c("variable_name", "variable_level"))

groupNameCol

groupAsColumn

Whether to display the group labels as a column (TRUE) or rows (FALSE).

groupNameAsColumn

groupOrder

Order in which to display group labels.

colsToMergeRows

Names of the columns to merge vertically when consecutive row cells have identical values. Alternatively, use "all_columns" to apply this merging to all columns, or use NULL to indicate no merging.

Value

gt object.

A gt table.

Examples

mockSummarisedResult() |>
  formatEstimateValue(decimals = c(integer = 0, numeric = 1)) |>
  formatHeader(header = c("Study strata", "strata_name", "strata_level"),
              includeHeaderName = FALSE) |>
  gtTable(
    style = list("header" = list(
      gt::cell_fill(color = "#d9d9d9"),
      gt::cell_text(weight = "bold")),
      "header_level" = list(gt::cell_fill(color = "#e1e1e1"),
                            gt::cell_text(weight = "bold")),
      "column_name" = list(gt::cell_text(weight = "bold")),
      "title" = list(gt::cell_text(weight = "bold"),
                     gt::cell_fill(color = "#c8c8c8")),
      "group_label" = gt::cell_fill(color = "#e1e1e1")),
    na = "--",
    title = "gtTable example",
    subtitle = NULL,
    caption = NULL,
    groupColumn = "group_level",
    groupAsColumn = FALSE,
    groupOrder = c("cohort1", "cohort2"),
    colsToMergeRows = "all_columns"
  )

  

gtTable example
    

      	
        Study strata
      
    

result_id
      	cdm_name
      	group_name
      	variable_name
      	variable_level
      	estimate_name
      	estimate_type
      	additional_name
      	additional_level
      	
        overall
      
      	
        age_group &&& sex
      
      	
        sex
      
      	
        age_group
      
    

overall
      	<40 &&& Male
      	>=40 &&& Male
      	<40 &&& Female
      	>=40 &&& Female
      	Male
      	Female
      	<40
      	>=40
    

cohort1
    

1
	mock
	cohort_name
	number subjects
	--
	count
	integer
	overall
	overall
	7,748,016
	1,514,956
	3,946,066
	9,349,984
	7,395,896
	6,356,681
	3,382,199
	9,624,163
	1,117,899

	
	
	age
	--
	mean
	numeric
	overall
	overall
	85.1
	50.1
	16.6
	36.1
	65.2
	87.4
	72.8
	41.2
	15.9

	
	
	
	
	sd
	numeric
	overall
	overall
	3.6
	3.9
	9.0
	2.8
	1.4
	4.5
	1.1
	8.8
	5.3

	
	
	Medications
	Amoxiciline
	count
	integer
	overall
	overall
	93,417
	8,444
	3,405
	31,096
	78,471
	28,040
	16,528
	43,359
	81,405

	
	
	
	
	percentage
	percentage
	overall
	overall
	2.21170505974442
	81.5023199422285
	30.3833421086892
	93.1795864831656
	54.043101449497
	25.4818350775167
	47.7587363682687
	70.4463991336524
	0.750555354170501

	
	
	
	Ibuprofen
	count
	integer
	overall
	overall
	17,487
	32,731
	61,475
	7,917
	51,906
	30,628
	67,140
	39,367
	74,785

	
	
	
	
	percentage
	percentage
	overall
	overall
	41.1679971730337
	74.6071676257998
	72.3450553836301
	61.4524137461558
	47.3980569280684
	57.8461039112881
	24.159372295253
	11.4378662081435
	29.8344632610679

cohort2
    

1
	mock
	cohort_name
	number subjects
	--
	count
	integer
	overall
	overall
	6,504,853
	9,835,116
	1,791,275
	1,379,704
	565,586
	5,463,676
	4,665,048
	8,963,784
	185,314

	
	
	age
	--
	mean
	numeric
	overall
	overall
	84.6
	88.4
	39.6
	3.8
	95.1
	78.4
	21.9
	33.1
	17.8

	
	
	
	
	sd
	numeric
	overall
	overall
	7.4
	7.2
	2.6
	6.7
	2.1
	6.9
	1.9
	7.8
	0.2

	
	
	Medications
	Amoxiciline
	count
	integer
	overall
	overall
	68,651
	77,731
	9,120
	74,262
	39,667
	29,603
	74,292
	46,651
	77,303

	
	
	
	
	percentage
	percentage
	overall
	overall
	77.2408777615055
	26.5793564263731
	97.104093618691
	19.0525120124221
	1.96958889719099
	61.2376979552209
	52.5731408968568
	63.4520042454824
	39.5447959192097

	
	
	
	Ibuprofen
	count
	integer
	overall
	overall
	13,893
	37,734
	64,138
	99,419
	36,856
	33,735
	91,369
	90,942
	23,838

	
	
	
	
	percentage
	percentage
	overall
	overall
	48.4119494911283
	56.7406318150461
	75.395310902968
	54.5425235992298
	75.9677961235866
	21.2967462604865
	83.9265023823828
	40.6697055324912
	87.3739800183102