Generate a formatted table from a <data.table>

This function combines the functionalities of formatEstimateValue(), formatEstimateName(), formatHeader(), and formatTable() into a single function. While it does not require the input table to be a <summarised_result>, it does expect specific fields to apply some formatting functionalities.

Usage

visTable(
  result,
  estimateName = character(),
  header = character(),
  groupColumn = character(),
  rename = character(),
  type = "gt",
  hide = character(),
  .options = list()
)

Arguments

result

A table to format.

estimateName

A named list of estimate names to join, sorted by computation order. Use <...> to indicate estimate names. This argument requires that the table has estimate_name and estimate_value columns.

header

A vector specifying the elements to include in the header. The order of elements matters, with the first being the topmost header. The vector elements can be column names or labels for overall headers. The table must contain an estimate_value column to pivot the headers.

groupColumn

Columns to use as group labels. By default, the name of the new group will be the tidy* column names separated by ";". To specify a custom group name, use a named list such as: list("newGroupName" = c("variable_name", "variable_level")).

*tidy: The tidy format applied to column names replaces "_" with a space and converts them to sentence case. Use rename to customise specific column names.

rename

A named vector to customise column names, e.g., c("Database name" = "cdm_name"). The function will rename all column names not specified here into a tidy* format.

type

The desired format of the output table. See tableType() for allowed options.

hide

Columns to drop from the output table.

.options

A named list with additional formatting options. visOmopResults::tableOptions() shows allowed arguments and their default values.

Value

A tibble, gt, or flextable object.

Examples

result <- mockSummarisedResult()
result |>
  visTable(
    estimateName = c("N%" = "<count> (<percentage>)",
                     "N" = "<count>",
                     "Mean (SD)" = "<mean> (<sd>)"),
    header = c("Estimate"),
    rename = c("Database name" = "cdm_name"),
    groupColumn = c("strata_name", "strata_level"),
    hide = c("additional_name", "additional_level", "estimate_type", "result_type")
  )

  

Result id
      	Database name
      	Group name
      	Group level
      	Variable name
      	Variable level
      	Estimate name
      	Estimate
    

overall; overall
    

1
	mock
	cohort_name
	cohort1
	number subjects
	-
	N
	7,713,282

	
	
	cohort2
	number subjects
	-
	N
	4,433,637

	
	
	cohort1
	age
	-
	Mean (SD)
	32.98 (0.89)

	
	
	cohort2
	age
	-
	Mean (SD)
	98.14 (7.24)

	
	
	cohort1
	Medications
	Amoxiciline
	N%
	18,154 (17.99)

	
	
	cohort2
	Medications
	Amoxiciline
	N%
	40,022 (94.84)

	
	
	cohort1
	Medications
	Ibuprofen
	N%
	56,552 (64.62)

	
	
	cohort2
	Medications
	Ibuprofen
	N%
	47,423 (24.78)

age_group &&& sex; <40 &&& Male
    

1
	mock
	cohort_name
	cohort1
	number subjects
	-
	N
	5,017,637

	
	
	cohort2
	number subjects
	-
	N
	3,061,836

	
	
	cohort1
	age
	-
	Mean (SD)
	39.95 (9.20)

	
	
	cohort2
	age
	-
	Mean (SD)
	78.17 (3.85)

	
	
	cohort1
	Medications
	Amoxiciline
	N%
	87,894 (79.50)

	
	
	cohort2
	Medications
	Amoxiciline
	N%
	34,734 (14.43)

	
	
	cohort1
	Medications
	Ibuprofen
	N%
	46,619 (92.13)

	
	
	cohort2
	Medications
	Ibuprofen
	N%
	57,504 (52.07)

age_group &&& sex; >=40 &&& Male
    

1
	mock
	cohort_name
	cohort1
	number subjects
	-
	N
	7,112,888

	
	
	cohort2
	number subjects
	-
	N
	2,693,460

	
	
	cohort1
	age
	-
	Mean (SD)
	69.92 (6.38)

	
	
	cohort2
	age
	-
	Mean (SD)
	60.55 (3.89)

	
	
	cohort1
	Medications
	Amoxiciline
	N%
	67,666 (52.92)

	
	
	cohort2
	Medications
	Amoxiciline
	N%
	8,135 (80.38)

	
	
	cohort1
	Medications
	Ibuprofen
	N%
	23,349 (28.88)

	
	
	cohort2
	Medications
	Ibuprofen
	N%
	21,033 (13.30)

age_group &&& sex; <40 &&& Female
    

1
	mock
	cohort_name
	cohort1
	number subjects
	-
	N
	918,984

	
	
	cohort2
	number subjects
	-
	N
	2,704,266

	
	
	cohort1
	age
	-
	Mean (SD)
	24.25 (4.17)

	
	
	cohort2
	age
	-
	Mean (SD)
	20.84 (1.30)

	
	
	cohort1
	Medications
	Amoxiciline
	N%
	50,744 (46.61)

	
	
	cohort2
	Medications
	Amoxiciline
	N%
	4,009 (29.43)

	
	
	cohort1
	Medications
	Ibuprofen
	N%
	74,081 (65.44)

	
	
	cohort2
	Medications
	Ibuprofen
	N%
	59,830 (81.60)

age_group &&& sex; >=40 &&& Female
    

1
	mock
	cohort_name
	cohort1
	number subjects
	-
	N
	1,059,088

	
	
	cohort2
	number subjects
	-
	N
	976,897

	
	
	cohort1
	age
	-
	Mean (SD)
	2.75 (7.02)

	
	
	cohort2
	age
	-
	Mean (SD)
	76.37 (2.85)

	
	
	cohort1
	Medications
	Amoxiciline
	N%
	48,141 (89.22)

	
	
	cohort2
	Medications
	Amoxiciline
	N%
	26,791 (38.14)

	
	
	cohort1
	Medications
	Ibuprofen
	N%
	29,534 (61.88)

	
	
	cohort2
	Medications
	Ibuprofen
	N%
	2,650 (30.78)

sex; Male
    

1
	mock
	cohort_name
	cohort1
	number subjects
	-
	N
	2,249,139

	
	
	cohort2
	number subjects
	-
	N
	9,920,600

	
	
	cohort1
	age
	-
	Mean (SD)
	7.01 (1.61)

	
	
	cohort2
	age
	-
	Mean (SD)
	81.88 (1.31)

	
	
	cohort1
	Medications
	Amoxiciline
	N%
	96,832 (74.20)

	
	
	cohort2
	Medications
	Amoxiciline
	N%
	48,458 (23.72)

	
	
	cohort1
	Medications
	Ibuprofen
	N%
	8,237 (64.43)

	
	
	cohort2
	Medications
	Ibuprofen
	N%
	85,237 (34.65)

sex; Female
    

1
	mock
	cohort_name
	cohort1
	number subjects
	-
	N
	4,176,636

	
	
	cohort2
	number subjects
	-
	N
	1,045,265

	
	
	cohort1
	age
	-
	Mean (SD)
	2.37 (3.78)

	
	
	cohort2
	age
	-
	Mean (SD)
	76.11 (4.74)

	
	
	cohort1
	Medications
	Amoxiciline
	N%
	24,795 (76.37)

	
	
	cohort2
	Medications
	Amoxiciline
	N%
	60,385 (34.61)

	
	
	cohort1
	Medications
	Ibuprofen
	N%
	27,436 (58.79)

	
	
	cohort2
	Medications
	Ibuprofen
	N%
	53,381 (32.44)

age_group; <40
    

1
	mock
	cohort_name
	cohort1
	number subjects
	-
	N
	3,440,131

	
	
	cohort2
	number subjects
	-
	N
	4,344,885

	
	
	cohort1
	age
	-
	Mean (SD)
	20.66 (9.75)

	
	
	cohort2
	age
	-
	Mean (SD)
	82.86 (9.80)

	
	
	cohort1
	Medications
	Amoxiciline
	N%
	13,281 (84.41)

	
	
	cohort2
	Medications
	Amoxiciline
	N%
	21,268 (42.90)

	
	
	cohort1
	Medications
	Ibuprofen
	N%
	93,106 (95.90)

	
	
	cohort2
	Medications
	Ibuprofen
	N%
	9,865 (83.00)

age_group; >=40
    

1
	mock
	cohort_name
	cohort1
	number subjects
	-
	N
	6,642,060

	
	
	cohort2
	number subjects
	-
	N
	6,846,876

	
	
	cohort1
	age
	-
	Mean (SD)
	69.89 (4.04)

	
	
	cohort2
	age
	-
	Mean (SD)
	0.85 (1.68)

	
	
	cohort1
	Medications
	Amoxiciline
	N%
	6,827 (70.52)

	
	
	cohort2
	Medications
	Amoxiciline
	N%
	8,016 (59.99)

	
	
	cohort1
	Medications
	Ibuprofen
	N%
	88,141 (16.80)

	
	
	cohort2
	Medications
	Ibuprofen
	N%
	69,738 (16.98)